fix image scaling, added menu bg

+++ CheckersEngine/Src/ECS/TransformComponent.h
@@ -10,7 +10,7 @@ public:
 	Vector2D velocity;
 	int height = 32;
 	int width = 32;
-	int scale = 1;
+	float scale = 1;
 
 
 	int speed = 3;
@@ -34,7 +34,7 @@ public:
 		position.y = y;
 	}
 
-	TransformComponent(float x, float y, int h, int w, int sc)
+	TransformComponent(float x, float y, int h, int w, float sc)
 	{
 		position.x = x;
 		position.y = y;
